+++ completions/sidenote.bash
@@ -28,7 +28,7 @@ _sidenote_path()
 
 _sidenote()
 {
-    local -r cmds=(cat edit exec import init ls path rm serve show)
+    local -r cmds=(cat completion edit exec import init ls path rm serve show)
     local cur prev
     _get_comp_words_by_ref cur prev
 
@@ -78,6 +78,16 @@ _sidenote()
             ;;
         *)
             _sidenote_path "${notes}"
+            ;;
+        esac
+        ;;
+    completion)
+        case "${cur}" in
+        -*)
+            COMPREPLY=($(compgen -W '-h' -- "${cur}"))
+            ;;
+        *)
+            COMPREPLY=($(compgen -W 'bash' -- "${cur}"))
             ;;
         esac
         ;;